[build] Fix three MSVC build issues in #3963 base
network_graph.cpp: add explicit unsigned-int casts for iterator difference_type when constructing execution-order tuple; MSVC 19.44 rejects the implicit narrowing conversion. mem_allocator.cpp: include windows.h before sysinfoapi.h so winnt.h sees the target-architecture macros (_M_X64) before it errors. causal_conv1d_layer: add CAUSAL_CONV1D_BUILDING_DLL guard and dllexport annotations; expose factory function unconditionally so the MSVC linker generates the import .lib needed by nntr_causallm.exe.

+++ nntrainer/graph/network_graph.cpp
@@ -115,8 +115,11 @@ void NetworkGraph::setExecutionOrder() {
       backward_order++;
     auto apply_gradient_order = backward_order++;
 
-    node->setExecutionOrder({forward_order, calc_gradient_order,
-                             calc_derivative_order, apply_gradient_order});
+    node->setExecutionOrder(
+      {static_cast<unsigned int>(forward_order),
+       static_cast<unsigned int>(calc_gradient_order),
+       static_cast<unsigned int>(calc_derivative_order),
+       static_cast<unsigned int>(apply_gradient_order)});
   }
 
   /**
